+++ changelog.md
@@ -1,6 +1,6 @@
 # ACEmulator Change Log
 
-### 2017-04-11
+### 2017-04-12
 [OptimShi]
 * Changed DatDatabase.AllFiles to a Dictionary<uint objectid, DatFile> (was simply a list List<DatFile> previously)
 * Added a DatReader class for reading from the a Dat file. Only confirmed for client_portal.dat.
@@ -10,8 +10,12 @@
 
 ### 2017-04-10
 [Ripley]
+* Added MovementEvent Action Queue
 * Changed GetNextSequence back to GetCurrentSequence for SequenceType.ObjectInstance
 * Animations work again!
+
+[Lidefeath]
+* Fix GameActionType.QueryHealth to work across landblocks
 
 ### 2017-04-09
 [Ripley]
@@ -185,3 +189,5 @@
 ### 2017-03-25
 [Mogwai]
 * weenies weenies everywhere. object structure pass 1
+
+